requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
praw==7.7.1
prawcore==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
